+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### Fixed
 
+- fixed 'Send'-button being functional for empty input
+
 ## [0.6.0] - 2025-03-22
 
 ### Added